cat Event.h Event.cxx MainEvent.cxx > MainEvent.C
